+++ common-core/change-log.md
@@ -13,6 +13,7 @@
 9. ClassUtils加入getPrimitiveType方法，forName方法支持基本类型(primitive type)
 11. 加入TreeUtils，可以从任意父子关系对象集合构建树形结构
 12. ClassUtils加入isRecord用于在jdk8的源码判断jdk16的record
+13. Lang的toString方法加入数组的支持（与Collection相同逻辑）
 
 # 1.14.0 2025-02-13